pyproject.toml
[tool.black]
extend-exclude = ""

[tool.pylint]
extension-pkg-allow-list = ["pygame"]

[tool.ruff]
extend-exclude = []
extend-ignore = ["E203"]
ignore-init-module-imports = true
select = ["E", "F", "I", "B", "C4", "SIM", "N", "FLY", "UP"]

[tool.ruff.flake8-bugbear]
extend-immutable-calls = []

[tool.ruff.isort]
force-single-line = true

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[project]
name = "Pygskin"
dynamic = ["version"]
dependencies = [
  "pygame-ce==2.5.2",
]
requires-python = ">=3.12"
authors = [
  {name = "Andy Driver", email = "andy@pagezero.net"},
]
maintainers = [
  {name = "Andy Driver", email = "andy@pagezero.net"},
]
readme = "README.md"
license = {text = "MIT"}
keywords = ["pygame", "game", "framework"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Libraries",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]

[project.urls]
homepage = "https://github.com/andyhd/pygskin"
documentation = "https://github.com/andyhd/pygskin/blob/main/README.md"
repository = "https://github.com/andyhd/pygskin"
issues = "https://github.com/andyhd/pygskin/issues"

